[2019 hw02]remove bottom node to avoid bashwash
- due to removal of bottom node, to track whether or not
the set is connected to bottom, I added an boolean array
connectedToBottom, if a disjoin set contains node in the bottom,
the array will contain true in the root index. To preserve the
infomation, union operation is wraped, so every time we union two set,
the connected to bottom infomation is preserved by or operation.

see link: https://drkbl.com/posts/quote-avoid-backwash-in-percolation/